Comment out more sdl_gfx code, and update SetColors usage
SDL gfx is more used than I first saw. At this point I wanted to add it as a
dependancy, but it turns out that the author has changed the api to work with
renderer instead of surfaces, so tough luck. I mailed him as I think the old
api should be kept, but I'm going on with that commented out anyway.

Also, updated SetColors to SetPaletteColors.