Update all the screen related stuff.
I tried to substitute all the 'screen' related code with the new SDL2 renderer
stuff. I keep having a 320x200 surface which is blitted into, and this ends up
in a streaming texture which is then rendered. I try to make use of the render
logic size, but of course I could not test anything of this.