Fix OptionInfo unintended 'overloading'
In SDL SDLKey was an enum, but in SDL2 SDL_Keycode (the substitute) is an int,
so there were 2 OptionInfo ctor with the same footprint. Create an inherited
class for the key one.

Also, don't try to do anything with mouse grab, as this changed in sdl2. I
still have to look at it.