Implement a dialog for the user to choose game's directories.
The dialog used is aFileDialog (https://code.google.com/p/afiledialog/) with minor changes. The sources are added to the main source tree to ease compilation. Some native glue code also had to be added to be able to set chosen directories without the need to restart the whole app.