Add placeholders for translation paths.
Yeah, the translations system is quite stupid for now: it merges files that you (should) get from Transifex with your own translations (in my case it's just a bunch of Android-specific strings) and then creates a latest.zip file containing the merged translations.
It's quite error-prone, but it should work for now. At least the build process is now pretty simple.