Don't rely solely on SDL_GetMouseState() for registering touches.
Since we're now using a new SDL revision, we're separating mouse and touch events
completely. This means better support for actual mice (near-desktop experience),
but this also means we can't rely on SDL_GetMouseState() reporting a finger on
the screen. In places where SDL_GetMouseState() was used (mostly to catch missed
mouseup events), there's now a function that checks if there's a mouse button pressed
or a finger is on the screen.

It might defeat the purpose of separating mouse and touch somewhat, but I think it's
for the best.