Re-initialize the game after directory change.
Since the inclusion of FileMap file accessing method, the old way of just
substituting new folders will not work - you have to re-scan your folders
again. Since these re-scanning functions are declared static (and for a
good reason, too), the easiest way to re-scan the folders is to
re-initialize the game (which is, to be fair, a much better solution than
the one before, since it should reload configuration files as well).