Fix possible dangling pointer issue.
Action class does not copy the SDL_Event structure, it only grabs the
pointer to it. The SDL_Event structure that was being created in
Inventory::mousePress was invalidated as soon as it got out of scope,
making the pointer to it invalid. What's even worse, in
Inventory::longPressAction it was being used to write information, which
could screw up things big time. Anyway, the SDL_Event is now stored as a
part of the State, which is still a hack, but so is long pressing.